[MediaStatusColumn] Add custom file for media status icons with colors
